hf_space/requirements.txt
flask==3.1.3
flask-cors==6.0.2
tensorflow==2.21.0
numpy
opencv-python-headless
pillow
